requirements.txt
libvirt-python>=1.3.1
psutil
